+++ src/ClientHandle.h
@@ -225,7 +225,7 @@ public:
 	*/
 	bool HandleLogin(int a_ProtocolVersion, const AString & a_Username);
 	
-	void SendData(const char * a_Data, int a_Size);
+	void SendData(const char * a_Data, size_t a_Size);
 	
 	/** Called when the player moves into a different world; queues sreaming the new chunks */
 	void MoveToWorld(cWorld & a_World, bool a_SendRespawnPacket);
